View archived pages
Before it just fired up the HTML, the JS, and the CSS from the archived
page. So if the page links CSS or JS using an absolute link, like
hosted bootstrap or something, that came in and screwed up the page.
Now, previews are images and if you want to see the whole page you can
hit “View raw”.